libgit2 1.5.0 (#1)
I added libgit2's Git repository as a submodule of this repository. I
checked out the source code for the libgit2 1.5.0 release.

I added the ios-cmake repository a submodule also and checked out the
4.3.0 release of the toolset for building libgit2 for macOS and iOS.

I created the build.sh program that automates building libgit2 for each
of the supported platforms and produces the XCFramework for linking
libgit2 to applications or other frameworks.